refactor(bundler): rename create-federation-tsconfig to update-federation-tsconfig
The module has not created anything since the schematic took over scaffolding;
it replaces `files` in a tsconfig that must already exist. The file name now
matches its export and contrasts with the schematic's generate- step.
